Move trace filters into the search
