+++ apps/web/src/routes/(app)/traces/+page.svelte
@@ -4,15 +4,17 @@
 	import type { OverlayScrollbars } from 'overlayscrollbars';
 	import { tick } from 'svelte';
 
+	import { goto } from '$app/navigation';
 	import { page } from '$app/state';
+	import { isTraceId } from 'api/schemas';
 	import TimeRangePicker from '$lib/components/search/TimeRangePicker.svelte';
-	import TraceFilters from '$lib/components/trace/TraceFilters.svelte';
 	import TraceHeatmapPanel from '$lib/components/trace/TraceHeatmapPanel.svelte';
 	import TraceList from '$lib/components/trace/TraceList.svelte';
 	import PanelError from '$lib/components/ui/PanelError.svelte';
 	import { TraceExplorerStore } from '$lib/stores/trace-explorer.svelte';
 	import { readLastIndex, writeLastIndex } from '$lib/utils/last-index';
 	import { OS_SCROLLBAR_BOTH_AXES_OPTIONS } from '$lib/utils/scrollbars';
+	import { traceDetailHref } from '$lib/utils/trace-params';
 
 	const SCROLL_TRIGGER_PX = 1500;
 
@@ -45,6 +47,17 @@
 		writeLastIndex(indexId);
 	}
 
+	let traceIdInput = $state('');
+	const traceIdValid = $derived(isTraceId(traceIdInput.trim().toLowerCase()));
+
+	function jumpToTrace(event: SubmitEvent): void {
+		event.preventDefault();
+		const id = traceIdInput.trim().toLowerCase();
+		if (!isTraceId(id)) return;
+		traceIdInput = '';
+		void goto(traceDetailHref(id, { index: selectedIndex, returnTo: page.url }));
+	}
+
 	function maybeFillViewport(os = osRef?.osInstance()): void {
 		if (store.traces.length === 0 || !store.hasMore) return;
 		const viewport = os?.elements().viewport;
@@ -65,7 +78,7 @@
 		aria-label="Trace search"
 	>
 		<select
-			class="select select-sm w-48 font-mono text-xs"
+			class="select select-sm w-44 font-mono text-xs"
 			value={selectedIndex}
 			onchange={(e) => selectIndex(e.currentTarget.value)}
 			aria-label="Logs index for trace links"
@@ -75,7 +88,36 @@
 			{/each}
 		</select>
 
-		<div class="flex-1"></div>
+		<select
+			class="select select-sm w-44"
+			value={store.params.service ?? ''}
+			onchange={(e) => store.navigate({ service: e.currentTarget.value || null }, { push: true })}
+			aria-label="Service"
+		>
+			<option value="">All services</option>
+			<!-- The roster is windowed, so the URL's service can be absent from it — without this the select
+			     renders blank and the active filter looks unset. -->
+			{#if store.params.service !== null && !store.services.includes(store.params.service)}
+				<option value={store.params.service}>{store.params.service}</option>
+			{/if}
+			{#each store.services as service (service)}
+				<option value={service}>{service}</option>
+			{/each}
+		</select>
+
+		<form class="min-w-0 flex-1" onsubmit={jumpToTrace}>
+			<input
+				type="text"
+				class="input input-sm validator w-full font-mono"
+				placeholder="Go to trace ID…"
+				bind:value={traceIdInput}
+				aria-label="Trace ID"
+				aria-invalid={traceIdInput !== '' && !traceIdValid}
+				title={traceIdInput !== '' && !traceIdValid
+					? 'Expected 32 hexadecimal characters'
+					: undefined}
+			/>
+		</form>
 
 		<TimeRangePicker
 			value={store.timeRange}
@@ -91,8 +133,6 @@
 			<RefreshCw class="h-3.5 w-3.5" aria-hidden="true" />
 		</button>
 	</div>
-
-	<TraceFilters {store} logIndexId={selectedIndex} />
 
 	<TraceHeatmapPanel
 		data={store.heatmap}
